Improve functionality of the card modal description section.
