+++ amma-track-app/src/redux/slices/cardModalSlice.js
@@ -11,16 +11,20 @@ export const cardModalSlice = createSlice({
     addDescription: (state, description) => {
       const desc = state.descriptions.find(card => card.cardID === description.payload.cardID);
 
-      if (!desc)
+      if (desc && description.payload.description === '<p><br></p>') {
+        state.descriptions = state.descriptions.filter(card => card.cardID !== description.payload.cardID);
+        return;
+      }
+
+      if (!desc) {
         state.descriptions = [...state.descriptions,  description.payload];
-      else {
+      } else {
         desc.description = description.payload.description
       }
       
     },
   },
 })
-
 
 export const { addDescription } = cardModalSlice.actions
